Add Redis helper to wait for rate limit expiry
