+++ feature.env
@@ -10,7 +10,10 @@ ENVIRONMENT_NAME=dev
 
 ORCHESTRATION_API_URL=http://localhost:9091/orchestration
 
-ADD_PRISONER_RATE_LIMIT_BOOKER_MAX_REQUESTS=2
-ADD_PRISONER_RATE_LIMIT_PRISONER_MAX_REQUESTS=3
+ADD_PRISONER_RATE_LIMIT_BOOKER_MAX_REQUESTS=3
+ADD_PRISONER_RATE_LIMIT_BOOKER_WINDOW_SECS=4
+
+ADD_PRISONER_RATE_LIMIT_PRISONER_MAX_REQUESTS=4
+ADD_PRISONER_RATE_LIMIT_PRISONER_WINDOW_SECS=3
 
 FEATURE_ADD_PRISONER_ENABLED=true
