--- integration_tests/redis/redisCache.ts
@@ -1,17 +1,0 @@
-import { createRedisClient } from '../../server/data/redisClient'
-
-const clearRateLimits = async () => {
-  const client = createRedisClient()
-  await client.connect()
-
-  const keys = await client.keys('rateLimit:*')
-
-  if (keys.length > 0) {
-    await client.del(keys)
-  }
-
-  await client.quit()
-  return Promise.resolve(null)
-}
-
-export default { clearRateLimits }
